+++ lib/model/device.dart
@@ -7,9 +7,10 @@ final Device device = Device._private();
 
 class Device {
   late final String _packageName, _packageVersion, _userAgent;
-  late final bool _canAuthenticateUser, _canPinShortcut;
+  late final bool _canAuthenticateUser, _canPinShortcut, _showPinShortcutFeedback;
   late final bool _canRenderFlagEmojis, _canRenderSubdivisionFlagEmojis, _canRequestManageMedia, _canSetLockScreenWallpaper;
-  late final bool _hasGeocoder, _isDynamicColorAvailable, _isTelevision, _showPinShortcutFeedback, _supportEdgeToEdgeUIMode, _supportPictureInPicture;
+  late final bool _hasGeocoder, _isDynamicColorAvailable, _supportEdgeToEdgeUIMode, _supportPictureInPicture;
+  late final bool _isPhysicalDevice, _isTelevision;
 
   String get packageName => _packageName;
 
@@ -33,6 +34,8 @@ class Device {
 
   bool get isDynamicColorAvailable => _isDynamicColorAvailable;
 
+  bool get isPhysicalDevice => _isPhysicalDevice;
+
   bool get isTelevision => _isTelevision;
 
   bool get showPinShortcutFeedback => _showPinShortcutFeedback;
@@ -50,6 +53,7 @@ class Device {
     _userAgent = '$_packageName/$_packageVersion';
 
     final androidInfo = await DeviceInfoPlugin().androidInfo;
+    _isPhysicalDevice = androidInfo.isPhysicalDevice;
     _isTelevision = androidInfo.systemFeatures.contains('android.software.leanback');
 
     final auth = LocalAuthentication();