+++ lib/model/settings/settings.dart
@@ -71,6 +71,7 @@ class Settings with ChangeNotifier, SettingsAccess, SearchSettings, AppSettings,
     _unregister();
     _register(monitorPlatformSettings);
     initAppSettings();
+    await sanitize();
   }
 
   void _unregister() {
@@ -184,6 +185,11 @@ class Settings with ChangeNotifier, SettingsAccess, SearchSettings, AppSettings,
       set(SettingKeys.videoBackgroundModeKey, null);
     }
     collectionBurstPatterns = collectionBurstPatterns.where(BurstPatterns.options.contains).toList();
+
+    // emulator
+    if (!device.isPhysicalDevice) {
+      set(SettingKeys.videoHardwareAccelerationKey, VideoHardwareAcceleration.disabled);
+    }
   }
 
   // tag editor
@@ -383,7 +389,6 @@ class Settings with ChangeNotifier, SettingsAccess, SearchSettings, AppSettings,
             case SettingKeys.viewerGestureSideTapNextKey:
             case SettingKeys.viewerUseCutoutKey:
             case SettingKeys.enableMotionPhotoAutoPlayKey:
-            case SettingKeys.enableVideoHardwareAccelerationKey:
             case SettingKeys.videoGestureDoubleTapTogglePlayKey:
             case SettingKeys.videoGestureSideDoubleTapSeekKey:
             case SettingKeys.videoGestureVerticalDragBrightnessVolumeKey:
@@ -426,6 +431,7 @@ class Settings with ChangeNotifier, SettingsAccess, SearchSettings, AppSettings,
             case SettingKeys.imageBackgroundKey:
             case SettingKeys.videoAutoPlayModeKey:
             case SettingKeys.videoBackgroundModeKey:
+            case SettingKeys.videoHardwareAccelerationKey:
             case SettingKeys.videoLoopModeKey:
             case SettingKeys.videoResumptionModeKey:
             case SettingKeys.subtitleTextAlignmentKey: